djangof/requirements.txt
asgiref==3.3.4
botocore==1.12.96
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-utils-six==2.0
djangorestframework==3.12.4
docutils==0.14
gunicorn==20.1.0
jmespath==0.9.3
pillow==2.4.0
psycopg2==2.8.6
python-dateutil==2.8.0
pytz==2018.5
s3transfer==0.2.0
six==1.12.0
sqlparse==0.4.1
urllib3==1.23
waitress==2.0.0
whitenoise==4.1.2
